refactor(account): split lib.rs into per-concern modules
`crates/account/src/lib.rs` had grown to 2124 lines — every id, credential,
chain walk and all 53 tests in one file. Split it along the crate's own
seams, with no behavior change: the public API is byte-identical, since
every item is re-exported flat from `lib.rs` and the modules stay private.

- `account.rs`    genesis version, `AccountGenesis`, member endorsements
- `root_key.rs`   handoff chain and `root_key_at_epoch`
- `device.rs`     `KemPublicKey`, `DeviceCert`, cert verification
- `revocation.rs` `DeviceRevocation` and its self-contained proof
- `pairing.rs`    pairing statement and confirmation code
- `domain.rs`     every signing domain in one place, plus `borsh_bytes`
- `error.rs`      `AccountError`

Domain separators are collected rather than left beside their users: the
property that matters is a property of the set — any two that collide let
a signature be replayed across purposes — so the distinctness test now
covers the whole crate instead of whatever module is in view.

Tests move out of the source files entirely, into one `src/tests/` tree —
`src/tests/<module>.rs` per module under test, cross-cutting borsh
round-trips in `src/tests/wire.rs`, shared fixtures in
`src/tests/support.rs`. Each file reaches its subject through `crate::`
paths, so reading a module is reading the model and nothing else.

Same 53 tests, all passing; `cargo check --workspace --all-targets`,
clippy `-D warnings`, `cargo fmt --check` and `cargo doc` are clean.
